rename queue_storage to session_storage
