+++ src/playback.rs
@@ -1,5 +1,5 @@
 pub mod events;
 pub mod interface;
 pub mod queue;
-pub mod queue_storage;
+pub mod session_storage;
 pub mod thread;
